+++ sdk/keyvault/Azure.Security.KeyVault.Keys/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added `CreateEcKeyOptions` class.
 - Added `CreateEcKey` and `CreateEcKeyAsync` methods to the `KeyClient` class.
+- Added constructor to `KeyVaultKeyIdentifier` to parse a `Uri`.
 
 ### Changed
 
@@ -20,6 +21,7 @@
 - Removed `Export` and `ExportAsync` methods from `KeyClient`.
 - Removed `Exportable` property from `KeyProperties`'.
 - Removed `KeyReleasePolicy` class and associated properties.
+- Removed `KeyVaultKeyIdentifier.Parse` and `KeyVaultKeyIdentifier.TryParse` in favor of the added constructor.
 
 ## 4.2.0-beta.3 (2020-11-12)